requirements.txt
python-decouple==3.8
requests==2.31.0
PyYAML